+++ course_memberships.go
@@ -1,8 +1,11 @@
 package blackboard_rest
 
 import (
+  "strings"
+
   "github.com/jbaxe2/blackboard.rest/api"
-  "github.com/jbaxe2/blackboard.rest/course_memberships"
+  courseMemberships "github.com/jbaxe2/blackboard.rest/course_memberships"
+  "github.com/jbaxe2/blackboard.rest/utils"
 )
 
 /**
@@ -12,20 +15,20 @@ import (
 type CourseMemberships interface {
   GetMembershipsForCourse (
     courseId string,
-  ) ([]course_memberships.Membership, error)
+  ) ([]courseMemberships.Membership, error)
 
-  GetMembershipsForUser (userId string) ([]course_memberships.Membership, error)
+  GetMembershipsForUser (userId string) ([]courseMemberships.Membership, error)
 
   GetMembership (
     courseId string, userId string,
-  ) (course_memberships.Membership, error)
+  ) (courseMemberships.Membership, error)
 
   UpdateMembership (
-    courseId string, userId string, membership course_memberships.Membership,
+    courseId string, userId string, membership courseMemberships.Membership,
   ) error
 
   CreateMembership (
-    courseId string, userId string, membership course_memberships.Membership,
+    courseId string, userId string, membership courseMemberships.Membership,
   ) error
 }
 
@@ -49,4 +52,49 @@ func NewCourseMemberships (service api.Service) CourseMemberships {
   return &_CourseMemberships {
     service: service,
   }
+}
+
+/**
+ * The [GetMembershipsForCourse] method obtains the collection of course
+ * memberships, based on the course's ID.
+ */
+func (memberships *_CourseMemberships) GetMembershipsForCourse (
+  courseId string,
+) ([]courseMemberships.Membership, error) {
+  return memberships._getMemberships (
+    string (api.CourseMemberships), "{courseId}", courseId,
+  )
+}
+
+/**
+ * The [GetMembershipsForCourse] method obtains the collection of course
+ * memberships, based on the user's ID.
+ */
+func (memberships *_CourseMemberships) GetMembershipsForUser (
+  userId string,
+) ([]courseMemberships.Membership, error) {
+  return memberships._getMemberships (
+    string (api.UserMemberships), "{userId}", userId,
+  )
+}
+
+/**
+ * The [_getMemberships] method obtains a collection of course memberships, for
+ * either the course or user based on the provided context type and context ID.
+ */
+func (memberships *_CourseMemberships) _getMemberships (
+  rawEndpoint, contextType, contextId string,
+) ([]courseMemberships.Membership, error) {
+  memberships.service.SetRequestOption ("expand", "user")
+
+  endpoint := strings.Replace (rawEndpoint, contextType, contextId, 1)
+  rawMemberships, err := memberships.service.Request (endpoint, "GET", nil, 1)
+
+  if nil != err {
+    return nil, err
+  }
+
+  return courseMemberships.NewMemberships (
+    utils.NormalizeRawResponse (rawMemberships["results"].([]interface{})),
+  ), nil
 }