+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 **blackboard.rest**
 
+## 0.2.9+7
+- Updated error response for requesting token.
+
+## 0.2.9+6
+- Library updates via minor refactoring.
+
+## 0.2.9+5
+- Added get course memberships for user.
+
+## 0.2.9+4
+- Added get course memberships for course.
+
 ## 0.2.9+3
 - Some fixes to more reliably handle API call responses.